AbstractDynamicsEffectInstanceModel
